(Engineering) Fixed compiler warnings
* Specifically, in backend and Jovial frontend

Issue #138

+++ src/frontend/Experimental_General_Language_Support/Tokens.h
@@ -84,7 +84,7 @@ public:
 
 private:
   std::vector<Token> tokens_;
-  unsigned long next_;
+  size_t next_;
 
   int getTokenElement(std::istream &, std::string &);
   int getTokenComment(std::istream &, std::string &);
